Upgraded WorldMap project to use GeoNode 2.7.x

+++ Dockerfile
@@ -1,28 +1,53 @@
-FROM geonode/django:geonode
-## Install Java
-RUN \
-  apt-get update -y && \
-  apt-get install -y openjdk-7-jdk && \
-  rm -rf /var/lib/apt/lists/*
+FROM python:2.7.14-stretch
+MAINTAINER GeoNode development team
 
-# Define commonly used JAVA_HOME variable
-ENV JAVA_HOME /usr/lib/jvm/java-7-openjdk-amd64
+RUN mkdir -p /usr/src/worldmap
 
-RUN mkdir -p /usr/src/app/
-COPY . /usr/src/app/
+WORKDIR /usr/src/worldmap
 
-## Install Ant
-ENV ANT_VERSION 1.9.4
-RUN cd && \
-    wget -q http://archive.apache.org/dist/ant/binaries/apache-ant-${ANT_VERSION}-bin.tar.gz && \
-    tar -xzf apache-ant-${ANT_VERSION}-bin.tar.gz && \
-    mv apache-ant-${ANT_VERSION} /opt/ant && \
-    rm apache-ant-${ANT_VERSION}-bin.tar.gz
+# This section is borrowed from the official Django image but adds GDAL and others
+RUN apt-get update && apt-get install -y \
+		gcc \
+		gettext \
+		postgresql-client libpq-dev \
+		sqlite3 \
+                python-gdal python-psycopg2 \
+                python-imaging python-lxml \
+                python-dev libgdal-dev \
+                python-ldap \
+                libmemcached-dev libsasl2-dev zlib1g-dev \
+                python-pylibmc \
+                uwsgi uwsgi-plugin-python \
+	--no-install-recommends && rm -rf /var/lib/apt/lists/*
 
-# Removing pre-compiled javascript.
-RUN rm -rf /usr/src/app/worldmap/static/worldmap_client/script
 
-WORKDIR /usr/src/app/src/geonode-client/
-RUN /opt/ant/bin/ant dist
+COPY wait-for-databases.sh /usr/bin/wait-for-databases
+RUN chmod +x /usr/bin/wait-for-databases
 
-WORKDIR /usr/src/app
+# Upgrade pip
+RUN pip install --upgrade pip
+
+# To understand the next section (the need for requirements.txt and setup.py)
+# Please read: https://packaging.python.org/requirements/
+
+# python-gdal does not seem to work, let's install manually the version that is
+# compatible with the provided libgdal-dev
+# superseded by pygdal
+#RUN pip install GDAL==2.1.3 --global-option=build_ext --global-option="-I/usr/include/gdal"
+RUN GDAL_VERSION=`gdal-config --version` \
+    && PYGDAL_VERSION="$(pip install pygdal==$GDAL_VERSION 2>&1 | grep -oP '(?<=: )(.*)(?=\))' | grep -oh $GDAL_VERSION\.[0-9])" \
+    && pip install pygdal==$PYGDAL_VERSION
+
+# fix for known bug in system-wide packages
+RUN ln -fs /usr/lib/python2.7/plat-x86_64-linux-gnu/_sysconfigdata*.py /usr/lib/python2.7/
+
+COPY . /usr/src/worldmap
+
+RUN chmod +x /usr/src/worldmap/tasks.py \
+    && chmod +x /usr/src/worldmap/entrypoint.sh
+
+# app-specific requirements
+RUN pip install --upgrade --no-cache-dir --src /usr/src -r requirements.txt
+RUN pip install --upgrade -e .
+
+ENTRYPOINT ["/usr/src/worldmap/entrypoint.sh"]
